# api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "直播互动引擎 - API 网关服务",
        "title": "Live Interact Engine - API Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/room/assign-role": {
            "post": {
                "description": "房间所有者为用户分配角色和权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "分配用户角色",
                "parameters": [
                    {
                        "description": "分配角色请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.AssignRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/create": {
            "post": {
                "description": "用户创建一个新房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "创建房间",
                "parameters": [
                    {
                        "description": "创建房间请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.CreateRoomReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "房间信息",
                        "schema": {
                            "$ref": "#/definitions/mapper.RoomResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/mute": {
            "post": {
                "description": "房间管理员禁言指定用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "禁言用户",
                "parameters": [
                    {
                        "description": "禁言用户请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.MuteUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/remove-role": {
            "post": {
                "description": "房间所有者移除用户权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "移除用户权限",
                "parameters": [
                    {
                        "description": "移除权限请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.RemoveRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/unmute": {
            "post": {
                "description": "房间管理员解除用户禁言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "解除禁言",
                "parameters": [
                    {
                        "description": "解除禁言请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.UnmuteUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/{room_id}": {
            "get": {
                "description": "根据房间ID获取房间详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "获取房间信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "房间信息",
                        "schema": {
                            "$ref": "#/definitions/mapper.RoomResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "房间不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/{room_id}/mute-list": {
            "get": {
                "description": "获取房间中被禁言的用户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "获取禁言列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "分页偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "分页大小",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "禁言列表",
                        "schema": {
                            "$ref": "#/definitions/mapper.GetMuteListResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/{room_id}/user/{user_id}/mute-info": {
            "get": {
                "description": "获取用户在房间的禁言详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "获取禁言信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "禁言信息",
                        "schema": {
                            "$ref": "#/definitions/mapper.MuteResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/{room_id}/user/{user_id}/mute-status": {
            "get": {
                "description": "检查用户是否在房间中被禁言",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "检查禁言状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "禁言状态",
                        "schema": {
                            "$ref": "#/definitions/mapper.IsMutedResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/{room_id}/user/{user_id}/permission/{permission}": {
            "get": {
                "description": "检查用户是否拥有指定权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "检查用户权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "权限值",
                        "name": "permission",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "权限检查结果",
                        "schema": {
                            "$ref": "#/definitions/mapper.CheckPermissionResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/room/{room_id}/user/{user_id}/role": {
            "get": {
                "description": "获取用户在指定房间的角色和权限信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "获取用户角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户角色信息",
                        "schema": {
                            "$ref": "#/definitions/mapper.UserRoomRoleResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "mapper.AssignRoleReq": {
            "type": "object",
            "required": [
                "role",
                "room_id",
                "user_id"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "administrator",
                        "vip"
                    ]
                },
                "room_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "mapper.CheckPermissionResp": {
            "type": "object",
            "properties": {
                "has_permission": {
                    "type": "boolean"
                }
            }
        },
        "mapper.CreateRoomReq": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "title": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                }
            }
        },
...
        "mapper.GetMuteListResp": {
            "type": "object",
            "properties": {
                "mutes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mapper.MuteResp"
                    }
                }
            }
        },
...
        "mapper.IsMutedResp": {
            "type": "object",
            "properties": {
                "is_muted": {
                    "type": "boolean"
                }
            }
        },
...
        "mapper.MuteResp": {
            "type": "object",
            "properties": {
                "admin_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "integer"
                },
                "muted_at": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "mapper.MuteUserReq": {
            "type": "object",
            "required": [
                "duration",
                "room_id",
                "user_id"
            ],
            "properties": {
                "duration": {
                    "type": "integer",
                    "minimum": 1
                },
                "reason": {
                    "type": "string",
                    "maxLength": 200
                },
                "room_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "mapper.RemoveRoleReq": {
            "type": "object",
            "required": [
                "room_id",
                "user_id"
            ],
            "properties": {
                "room_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "mapper.RoomResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "owner_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
        "mapper.UnmuteUserReq": {
            "type": "object",
            "required": [
                "room_id",
                "user_id"
            ],
            "properties": {
                "room_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "mapper.UserRoomRoleResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}